Extract logging severity
